+++ packages/kit-headless/src/components/checkbox/index.ts
@@ -1,3 +1,2 @@
-export { CheckboxRoot as Root } from './checkbox';
+export { CheckboxRoot as Root } from './checkbox-root';
 export { CheckboxIndicator as Indicator } from './checkbox-indicator';
-export * from './context-id';